requirements.txt
Django>=3.0,<5.0
firebase-admin
google-cloud-firestore
requests
